requirements.txt
streamlit
exa_py
google.generativeai
